Enable containers in the dewars table to wrap
The expectedContainers field was a long string without any spaces. This
was causing the column in the dewars table to be extremely long and
squishing the other columns. In this change we replace commas in the
expectedContainers field with spaces to enable line breaks to be
inserted. This means the containers column is no longer editable so we
may need to revisit this change if this feature is needed.