Enable dewars to be labelled missing
Sometimes users will take their dewars without scanning them out. In
order to track this and inform the users we cannot locate their dewar we
need to be able to flag dewars as missing. This change adds that
functionality to the dewars table.

We also add tooltips to the buttons in the dewars table to make the
function clear.