Make dewar reducer set addedTime if it is not given
When dewars are added by posting to the pucktracker-server the action
payload may not contain addedTime (eg when the barcode is scanned). In
this case clients should set addedTime to be the current time. The
server will also do this when saving to the database. The times will be
slightly out but this should not cause any issues.