Add indication when dewars need to be filled
Staff need to easily see which dewars need to be filled. To achieve this
we modify the On Site dewars table to highlight dewars which haven't been filled
in the last 3 days.

This change also removes the EPN from the dewar selection field on the
puck transfers page as the new dewar IDs contain this information.