Show dewar fill times relative to now
To make the dewar table last filled column easier to read we express the
times as relative to now (eg "3 hours ago").

This change also addresses a bug where time strings sent via socket.io are
not correctly parsed into Date objects. We currently only parse the
times in the SET_DEWARS and ADD_DEWAR actions as we are investigating
better ways that will apply to all actions.