Highlight rows where dewar is overdue for departure
Staff need to be made aware when a dewar should have departed the
facility. This change highlights the row of said dewars. We also change
the colour of the filled cell or the experiment end time cell to signal
the source of the warning.

We also update webpack and remove an unused html class from the
expected containers cell.