Enable submitting forms with Enter key without reloading page
The "add dewar" and "add puck" forms allowed submission via the Enter
key however this also caused the page to reload. This is an unwanted
disruption to the user experience. This change prevents the page reload
by disable the default behaviour of the form submission which is to do a
GET request.

We also disable form submission in the TypeaheadInput component as that
could cause unwanted page reloads on the puck transfer page.