When adding pucks to a dewar, enable submission via Enter key
Users expect that hitting enter while inside the "add puck" text box
should cause the puck to be added to the dewar. This was not happening
as the input box was initially being used to filter pucks and the user
was expected to make a selection from the options below. This change
enables the Enter key to submit the puck by assuming the first option in
the filtered list is the desired puck.