Clear pucks' dewar field when dewars are deleted
Prior to this change, when a dewar was deleted any pucks in that dewar
remained labelled as being inside that dewar, potentially leading to
confusion and preventing pucks from being reassigned to newly shipped
dewars. This change fixes this by clearing the puck receptacle and
receptacleType fields when the dewar is deleted.