Simplify config by removing development / production variants
Prior to this change the config.js file could contain both development
and production variants to be selected using the NODE_ENV. However this
does not work for client-side js where NODE_ENV is unavailable. We now
only allow a single configuration and developers can have two copies of
the codebase for development and production testing.