Make the expected containers cell editable
The expected containers cell was previously altered to replace commas
with spaces to allow the text to wrap. Since the expectedContainers
field in the database now has the containers separated by " | " the
text replacement is not necessary. This means we can make the cell use
the EditableCell class again so staff can edit the text directly inside
the table.