Show expected pucks in the dewar panel on the transfer page
To make it easier for staff to verify they are putting the right pucks
into a dewar we display the expected pucks in the dewar receptacle
component.

We also modify the DewarTypeReceptacle component to check the selected
dewar is actually valid. This is important as the selected dewar is
pulled from the url and so can be made invalid by user manipulation or
if an out of date url is accessed.